modularization of processInput
