+++ src/app/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    processInput: require('./processInput.js')
+}
